Share enum code between OpenAPI v2 and v3
