+++ adl/core/serialization/openapi/v2/schema.ts
@@ -1,31 +1,12 @@
 import { items, length, values } from '@azure-tools/linq';
-import { IntegerFormat, NumberFormat, StringFormat, v2, XMSEnumValue } from '@azure-tools/openapi';
+import { IntegerFormat, NumberFormat, StringFormat, v2 } from '@azure-tools/openapi';
 import { anonymous, isUsed, nameOf, unusedMembers, use, using } from '@azure-tools/sourcemap';
 import { Alias as GenericAlias } from '../../../model/alias';
 import { Identity } from '../../../model/name';
-import { Alias, ArraySchema, Constant, DictionarySchema, Enum, ExclusiveMaximumConstraint, ExclusiveMinimumConstraint, MaximumConstraint, MaximumElementsConstraint, MaximumPropertiesConstraint, MaxLengthConstraint, MinimumConstraint, MinimumElementsConstraint, MinimumPropertiesConstraint, MinLengthConstraint, MultipleOfConstraint, ObjectSchema, Property, ReadOnlyConstraint, RegularExpressionConstraint, Schema, ServerDefaultValue, UniqueElementsConstraint } from '../../../model/schema';
+import { Alias, ArraySchema, DictionarySchema, ExclusiveMaximumConstraint, ExclusiveMinimumConstraint, MaximumConstraint, MaximumElementsConstraint, MaximumPropertiesConstraint, MaxLengthConstraint, MinimumConstraint, MinimumElementsConstraint, MinimumPropertiesConstraint, MinLengthConstraint, MultipleOfConstraint, ObjectSchema, Property, ReadOnlyConstraint, RegularExpressionConstraint, Schema, ServerDefaultValue, UniqueElementsConstraint } from '../../../model/schema';
 import { isEnumSchema, isObjectSchema, push, singleOrDefault } from '../common';
-import { arrayProperties, commonProperties, numberProperties, objectProperties, processAnySchema, processBooleanSchema, processByteArraySchema, processCharSchema, processDateSchema, processDateTimeSchema, processDurationSchema, processFileSchema, processOdataSchema, processPasswordSchema, processTimeSchema, processUriSchema, processUuidSchema, stringProperties } from '../common/schema';
+import { arrayProperties, commonProperties, numberProperties, objectProperties, Options, processAnySchema, processBooleanSchema, processByteArraySchema, processCharSchema, processDateSchema, processDateTimeSchema, processDurationSchema, processEnumSchemaCommon, processFileSchema, processOdataSchema, processPasswordSchema, processTimeSchema, processUriSchema, processUuidSchema, stringProperties } from '../common/schema';
 import { Context } from './serializer';
-
-
-/** Schema processing options */
-type Options = Partial<{
-  /** this is an inline-declared anonymous schema; the name is not intended to be used as the final name */
-  isAnonymous: boolean;
-
-  /** processes the schema just as the target type, and not oneOf/allOf/anyOf/object/enum */
-  justTargetType: boolean;
-
-  /** note that this is a property declaration while processing this schema */
-  isProperty: boolean;
-
-  /** note that this is a parameter declaration while processing this schema */
-  isParameter: boolean;
-
-  forUnderlyingEnumType: boolean;
-}>;
-
 
 export async function* processInline(schema: v2.Schema | v2.SchemaReference | undefined, $: Context, options?: Options): AsyncGenerator<Schema> {
   if (schema) {
@@ -491,34 +472,8 @@ export async function* processObjectSchema(schema: v2.Schema, $: Context, option
 }
 
 export async function* processEnumSchema(schema: v2.Schema, $: Context): AsyncGenerator<Schema> {
-  const schemaEnum = use(schema.enum) ?? [];
-  const xmsEnum = use(schema['x-ms-enum']) ?? {};
-  const values: Array<XMSEnumValue> = xmsEnum.values ?? schemaEnum.map(v => ({ value: v }));
-
   // not using $.process here because we need to process a node that is already marked
   const type = await singleOrDefault(processSchema(schema, $, { forUnderlyingEnumType: true })) || $.api.schemas.Any;
-
-  const result = new Enum(type, {
-    name: xmsEnum.name || nameOf(schema),
-  });
-
-  result.sealed = !xmsEnum.modelAsString;
-  use(xmsEnum.modelAsString );
-
-  for (const each of values) {
-    const constant = new Constant(type, use(each.value), {
-      name: each.name,
-      description: each.description
-    });
-    result.values.push(constant);
-  }
-
-  // an enum with only one value is treated as single constant directly
-  if (result.values.length == 1) {
-    return yield result.values[0];
-  }
-
-  // yield the value as soon as possible so that if we start to recurse it's already in the cache.
-  yield result;
+  return yield* processEnumSchemaCommon(schema, $, type);
 }
